Futher codebase breakdown, and CONTRIBUTING.md
