+++ CONTRIBUTING.md
@@ -1,0 +1,29 @@
+# Contributing to create-bhvr
+
+First off, thank you for considering contributing to `create-bhvr`.
+
+## Code of Conduct
+
+This project and everyone participating in it is governed by the [Code of Conduct](CODE_OF_CONDUCT.md). By participating, you are expected to uphold this code.
+
+## How Can I Contribute?
+
+### Reporting Bugs
+
+If you find a bug, please make sure the bug has not already been reported by searching on GitHub under [Issues](https://github.com/stevedylandev/bhvr/issues). If you're unable to find an open issue addressing the problem, [open a new one](https://github.com/stevedylandev/bhvr/issues/new). Be sure to include a **title and clear description**, as much relevant information as possible, and a **code sample** or an **executable test case** demonstrating the expected behavior that is not occurring.
+
+### Suggesting Enhancements
+
+If you have an idea for an enhancement, please make sure the enhancement has not already been suggested by searching on GitHub under [Issues](https://github.com/stevedylandev/bhvr/issues). If you're unable to find an open issue addressing the suggestion, [open a new one](https://github.com/stevedylandev/bhvr/issues/new). Be sure to include a **title and clear description** of the enhancement you're suggesting.
+
+### Submitting a Pull Request
+
+1.  Fork the repository and create your branch from `main`.
+2.  Run `bun install` to install the dependencies.
+3.  Make your changes.
+4.  Run `bun run build` to make sure your changes build correctly.
+5.  Issue that pull request!
+
+## License
+
+By contributing, you agree that your contributions will be licensed under its MIT License.
