+++ src/lib/index.ts
@@ -1,4 +1,8 @@
 export * from "./create-project";
 export * from "./display-banner";
+export * from "./initialize-git";
+export * from "./install-dependencies";
 export * from "./patch-files-rpc";
+export * from "./prompt-for-options";
+export * from "./scaffold-template";
 export * from "./setup-biome";
