+++ src/types.ts
@@ -1,21 +1,22 @@
 export interface TemplateInfo {
-	branch: string;
-	description: string;
+  branch: string;
+  description: string;
 }
 
 export type ProjectOptions = {
-	yes?: boolean;
-	typescript?: boolean;
-	repo?: string;
-	template?: string;
-	branch?: string;
-	rpc?: boolean;
-	linter?: 'eslint' | 'biome';
+  projectName?: string;
+  yes?: boolean;
+  typescript?: boolean;
+  repo?: string;
+  template?: string;
+  branch?: string;
+  rpc?: boolean;
+  linter?: "eslint" | "biome";
 };
 
 export interface ProjectResult {
-	projectName: string;
-	gitInitialized: boolean;
-	dependenciesInstalled: boolean;
-	template: string;
+  projectName: string;
+  gitInitialized: boolean;
+  dependenciesInstalled: boolean;
+  template: string;
 }
